+++ firestore-stripe-invoices/CHANGELOG.md
@@ -1,3 +1,61 @@
+## Version 0.1.6 - 2021-02-11
+
+[feat] - Add `stripeInvoiceUrl` to the invoice object in Cloud Firestore. #132
+
+[feat] - Support invoices in various different currencies for the same email address. Note that this will create a separate customer object for each currency but with the same email address.
+
+[feat] - Add support for [`default_tax_rates`](https://stripe.com/docs/api/invoices/create#create_invoice-default_tax_rates) for the invoice and [`tax_rates`](https://stripe.com/docs/api/invoiceitems/create#create_invoiceitem-tax_rates) for the invoice items. Read more details about how to use tax rates in the [Stripe docs](https://stripe.com/docs/billing/taxes/tax-rates). #118
+
+```js
+await firestore
+  .collection("invoices")
+  .doc()
+  .set({
+    email: "testr@test.de",
+    default_tax_rates: ["txr_1HCkCjHYgolSBA35vh6cyHB5"],
+    items: [
+      {
+        amount: 1099,
+        currency: "usd",
+        description: "item 1",
+        tax_rates: ["txr_1HCshzHYgolSBA35WkPjzOOi"],
+      },
+      {
+        amount: 1250,
+        currency: "usd",
+        description: "item 2",
+      },
+    ],
+  });
+```
+
+[feat] - Add support for [`transfer_data`](https://stripe.com/docs/api/invoices/create#create_invoice-transfer_data) to send funds to a connected account upon successful payment.
+
+```js
+await firestore
+  .collection("invoices")
+  .doc()
+  .set({
+    transfer_data: {
+      destination: "acct_1234",
+      amount: 2114,
+    },
+    email: "testr@test.de",
+    items: [
+      {
+        amount: 1099,
+        currency: "usd",
+        description: "item 1",
+      },
+      {
+        amount: 1250,
+        currency: "usd",
+        description: "item 2",
+      },
+    ],
+  });
+```
+
 ## Version 0.1.5 - 2020-11-19
 
 [fixed] - Prevent users from specifying both email address and uid, only either one is allowed.